Check ID and visit specification for I2C2

Checks the input for I2C2

check_id_visit(y, id, visit)

Arguments

y
An n by p data matrix containing n vectorized image data with p voxels. Each row contains one observed image data at a particular visit for one subject. Each column contains image values for all subjects and visits at a particular voxel.

The rows are organized by subjects and then visits, EX) (Y11, Y12, Y21, Y22, ... , YI1 , YI2)

id
Vector of IDs, EX) c(1, 1, 2, 2, 3, 3, 4, 4, ... , I, I)

visit
Vector of visits, EX) (1, 2, 1, 2, 1, 2, ... , 1, 2)

Value

List of elements: reordered data, id, visit, I (number of ids), number of rows (n), and number of columns (p)

Examples

id = c(1:10, 10:1)
visit = rep(1:2, each = 10)
visit = as.character(visit)
n = length(id)
p = 100
y = matrix(rnorm(n * p), nrow = n, ncol = p)
check_id_visit(y =y, id = id, visit = visit)
#> $y
#>              [,1]        [,2]       [,3]       [,4]        [,5]        [,6]
#>  [1,] -1.62667268 -0.10176101 -0.6776146  0.5627448 -1.18394507  1.02816798
#>  [2,] -0.87215883 -0.30120868 -0.4812086  2.2205197 -1.65805086 -1.06262279
#>  [3,] -0.25647839 -0.25378053  0.6552276 -1.2463197 -0.07395583  1.12795361
#>  [4,] -0.68755543  0.23252515  0.7356907  0.6597641  0.66098669  0.24875801
#>  [5,]  1.10177950 -1.85374045 -0.4006375  0.3809222 -0.41635467 -1.28015460
#>  [6,] -0.37366156 -0.02881534  1.2161258  0.1588306  0.56317466 -0.91480448
#>  [7,]  0.75578151 -0.07794607 -0.3345566 -1.4304273 -0.19148234  1.12886823
#>  [8,] -1.26519802 -0.96765199  1.0663764  0.4402419 -0.63321301  0.62368124
#>  [9,] -0.23823356  0.96856634  1.3679540 -1.0484455  0.06954478 -0.46413453
#> [10,] -2.11906119  0.83447390  0.3401156  0.7166008  0.71157397 -1.19558030
#> [11,]  0.98744470  0.18492596  2.1377671 -0.2185036  1.15534832 -0.31576021
#> [12,] -0.67486594 -0.82866254 -0.8303227  0.7128943 -1.03994336 -0.03173974
#> [13,]  0.74139013 -1.37994358  0.5058193 -1.4899362  0.59495735  0.92429315
#> [14,] -2.18528684 -0.76254488 -0.4561763 -0.2226615  0.41937541  0.20288178
#> [15,]  0.08934727 -1.43551436  0.7863424  1.1727063 -1.41964511  0.07714472
#> [16,] -0.59631064  1.08655140 -1.2375945  0.5928281 -0.76180434 -0.48136561
#> [17,] -0.95494386  0.36208723 -0.9022119 -1.4798270 -1.60667725  1.03992361
#> [18,]  0.48297852 -0.65156299  0.2909875  1.5225863  1.22702839  0.95091897
#> [19,] -0.19515038 -1.75908675  0.5328970 -0.4303878  0.89292590  0.74188621
#> [20,]  0.92552126 -0.32454401 -0.6458943 -1.0516386  0.14816796  1.25554486
#>             [,7]        [,8]       [,9]       [,10]       [,11]       [,12]
#>  [1,] -0.3639822  1.33778247  0.1837639  0.69131734 -0.19366673  0.56061070
#>  [2,]  0.1019834 -1.46817333  0.2812456 -1.27464995  1.63089174 -0.35787991
#>  [3,] -1.2069949  0.08723477 -1.4351472  0.04636144 -1.29508364  1.32956476
#>  [4,]  0.5306499  0.95914119 -0.3052848  1.05262756  0.09731162  0.09258465
#>  [5,]  1.4292128 -0.39110421 -1.1373999  0.11302936  0.14188028 -0.27882350
#>  [6,]  1.0658793 -0.67731930  2.0682096  0.49582159  0.73351542 -0.12255013
#>  [7,]  0.6334359 -0.24986748 -0.4146453  0.99533187  1.26171505 -1.26673154
#>  [8,]  0.3724723 -0.43890931  0.6193027 -1.08081702 -0.05663631  0.50911870
#>  [9,] -1.9968156  1.15510475  0.1439343 -0.68115136 -0.43150033 -0.24914839
#> [10,]  1.4707523 -0.69940307 -0.1985061  0.87204699 -0.30760710  0.39515582
#> [11,] -0.6818322 -0.86472724  1.0620243 -1.27705725 -1.82271259  0.01798841
#> [12,] -1.9632525  0.52105453 -1.3603061  1.39027512  0.43242913  0.86208221
#> [13,] -0.4600555 -0.86667834 -0.5707939 -1.46869775  0.35254396  0.37707273
#> [14,] -1.2051854 -0.54940849  0.6435000  0.33087649 -0.28282162  0.74442373
#> [15,] -0.9830692 -2.32101703  1.2771814 -0.31347407 -1.34845144  0.79600856
#> [16,] -1.6753322  0.65316619  0.4255131  1.21551378 -0.32922472  1.60511404
#> [17,]  0.4953317  0.60883017  0.2282893 -1.70365949  0.70768832 -0.84067742
#> [18,]  0.9547864 -1.58000075  0.5488224 -1.09954301 -1.04115630 -1.34130996
#> [19,]  0.7258175  1.15000605 -0.3088131 -1.35051466 -0.41089094 -2.20547175
#> [20,]  0.6672987 -1.19959767  0.9598291 -1.10209368 -0.44604518 -1.12805599
#>             [,13]       [,14]      [,15]       [,16]       [,17]      [,18]
#>  [1,] -0.35965522  0.03102248  0.2668064 -1.40350302  0.02231080  2.1145213
#>  [2,] -2.43661560 -0.25732459  0.7668439 -0.44807823  0.38454106 -0.7589531
#>  [3,]  1.02857072 -0.33574670  1.3877243 -1.17604678  0.04446730 -1.0285674
#>  [4,]  0.11960089  0.07319578 -0.4455832 -0.23187451 -0.61667341  0.8316974
#>  [5,]  1.07789259 -0.26377616  0.1930796  0.51159647  0.48072117 -0.8867880
#>  [6,]  0.80634826 -1.21359356  0.3227550  0.29560157  1.61941870 -1.7366858
#>  [7,]  0.93178122 -0.35585246  0.5923166  1.31676526  0.63561986  1.2711460
#>  [8,]  0.98035352  0.04138848  0.1253846 -0.40443364 -0.72399509 -0.9684155
#>  [9,] -1.46079387 -0.48115480 -0.8299745  2.39291295 -2.00074239 -1.6050854
#> [10,]  0.01625547  1.18041020  0.7923495 -0.16927557  0.62347802 -1.2877952
#> [11,] -0.90607558  0.22273547  0.3925733 -0.06776941 -0.69065550  1.1222734
#> [12,] -2.32987955 -1.13956094 -0.6306855  0.83387906 -0.45879293 -1.3996571
#> [13,] -0.68034783  2.42991032  0.3848676 -0.20378507 -0.13064519  2.1584386
#> [14,]  0.47774044  0.06386348  0.5904095  0.95794490  0.28513338  0.7264094
#> [15,]  1.06316604  1.49629812  1.0510447  1.99643445 -0.34425178  0.4282466
#> [16,]  0.41027733  0.32751014  1.5025446  1.05534519  1.23544588  1.0946298
#> [17,] -0.69241315 -0.71720820  1.1557975  1.54321779 -0.05241282  1.2011787
#> [18,] -1.01290362  2.30017689  1.2736843  0.75701371  0.55479348  2.0120818
#> [19,] -1.13362827 -0.46705390 -1.0344380 -1.22833763  0.09612392  1.0316901
#> [20,] -1.09461543  0.66292345 -0.2544681  2.64448799  0.26640705  0.6538743
#>              [,19]       [,20]       [,21]       [,22]       [,23]       [,24]
#>  [1,]  2.155077001  1.89078004  1.19778677  1.51851513  0.03474609  1.06519498
#>  [2,] -0.612941179 -0.94099782  1.22951605  1.38848898 -0.74369442 -1.84401676
#>  [3,] -0.423258880 -0.14541053 -0.62423939  1.34509418 -1.29258503 -0.03208173
#>  [4,] -0.682838619 -1.95004570 -0.70483935  0.47314510  2.70063662  0.10610584
#>  [5,] -0.005649937 -0.63942372 -0.13286966 -0.12369153  1.16396750 -0.53995969
#>  [6,] -1.051996428 -1.95648670 -0.22865737 -0.90355250 -1.49346018 -0.61625270
#>  [7,]  0.400361729 -0.41117605 -0.01722092  0.93390245 -1.52338460 -1.92564357
#>  [8,] -1.449760536 -0.73835910  0.61204551 -0.08038236 -0.01917784 -1.46500068
#>  [9,] -0.365677802  0.34279188 -0.45962928  0.01949685 -2.51833505  1.00783138
#> [10,] -0.556624165  0.40617493  0.75226244  0.27421492  1.10636302  0.83869012
#> [11,] -0.429143835 -1.13324675  1.47293679  0.83085211 -0.70752923 -0.27923480
#> [12,]  0.730621937  2.18614406 -0.59312866  1.00679266 -0.08603859  1.32715824
#> [13,]  0.244697127  0.06607030  2.16931961 -0.30449467 -0.28743288 -1.24956586
#> [14,] -0.350470617  1.18175155  1.23305746 -0.75936325  1.75174562 -0.25500289
#> [15,]  2.219776249  0.03799977 -3.00143131 -0.05928128 -0.43530831  1.24849113
#> [16,] -0.528063683 -0.07945702  0.47489342  1.54674727 -0.72895427 -0.42810997
#> [17,]  0.314269808  0.92106475 -1.77198558  1.00739670 -0.34307877 -0.19221338
#> [18,]  0.415298593  1.24444646 -1.23407793 -0.02328030 -0.35994840  1.28961053
#> [19,] -0.563785075  2.13496564 -0.36483922 -1.37779509 -0.03931088  0.91197623
#> [20,] -1.485696158 -0.59982693  0.37453378 -1.00420273  0.88696599 -2.15826968
#>             [,25]       [,26]       [,27]       [,28]       [,29]       [,30]
#>  [1,] -1.61642285  1.65721450 -1.34984289 -0.58772242  0.36718999  0.95152023
#>  [2,]  0.02943674 -2.40000254 -1.77230491 -0.44316679 -0.53902163 -1.55245133
#>  [3,]  0.22767670  1.13122314  1.20200790 -1.28237031  0.69582170 -0.10557985
#>  [4,]  1.23996869  1.67843073  0.79188927  1.42636485 -0.63751255 -0.20414166
#>  [5,]  0.92121670 -1.40241063  1.29384272 -0.31213466  0.93344473 -0.37406927
#>  [6,]  1.69435049  0.62792225 -1.19344507  0.17993845  0.66031920 -1.21122382
#>  [7,] -0.33120344  0.15677046 -0.05959435 -0.69877530  1.49188908 -0.17623258
#>  [8,]  0.79349943 -0.29788176  0.45896221 -0.20747125 -0.58532696  1.39571912
#>  [9,]  1.28723403  0.86750860 -0.68506235 -2.64384977  1.15952299  0.27648331
#> [10,] -0.15943804  0.67769134  1.03472122  0.74151074 -0.28047404  0.31723058
#> [11,]  0.05726080 -0.03323516 -1.45343930  0.82502238  0.22679596  0.23633259
#> [12,]  0.17321454 -0.63324847 -0.58643431  3.54114028 -1.56982016 -1.35675471
#> [13,] -2.20393313  0.95406456  1.29412656  0.30692631 -1.29546518  2.06167416
#> [14,] -0.68000960 -0.11829945 -0.94142550  0.89039776 -0.41697042 -0.05474996
#> [15,] -0.52291035  1.13009673  1.17616909 -0.05803946 -0.06277869 -1.29005999
#> [16,]  0.86947504 -1.32980347 -1.07304684 -1.05271518 -1.18760009 -0.90197026
#> [17,]  1.08078531  0.39369731  0.86404508  0.92656941  0.99328294  0.57118945
#> [18,] -0.86829951  0.50409489 -1.38587675 -1.92333434  1.74401638  0.88469215
#> [19,] -0.24314134  0.37680833 -2.29673252 -0.21869186 -1.00593094  1.78539023
#> [20,] -0.90944656 -0.25174291 -1.39181001  0.99313350 -0.75098495  0.21294024
#>             [,31]      [,32]       [,33]       [,34]      [,35]      [,36]
#>  [1,]  0.51804967 -0.2869618 -0.91669760 -0.72968507  1.4054456 -0.2069272
#>  [2,] -0.82164842  0.2619207  1.06255745 -0.69084336 -1.6973907 -0.8924924
#>  [3,] -0.41884030  0.1312789 -0.56856448  0.05654772 -1.7013427 -0.3985970
#>  [4,] -0.08987210 -1.7317035 -0.34473014  0.45126628  0.5071799 -0.7738908
#>  [5,] -1.70945322 -1.0620840 -0.86582044  0.46782971  0.1922456 -1.1886198
#>  [6,]  2.39915487 -0.3366591  3.01290575 -1.61880052 -0.3362542 -0.6009867
#>  [7,]  2.48316267  0.7335810  0.82549391 -0.37264525 -0.7029568  0.4525654
#>  [8,] -0.56071976  0.6441241  0.26560942 -0.03119892 -1.6668793 -0.8395836
#>  [9,]  0.49803374  0.3226922 -0.80013228  0.21016169 -0.8085559  0.2075897
#> [10,] -0.04924635 -0.8068367  0.33386020  0.39099889  0.6162343 -0.4288354
#> [11,] -0.67386821 -0.3489798  0.97583009  1.28960831 -0.6381987 -1.2101224
#> [12,] -0.67685722 -0.2012890 -0.36172476  1.23840228  0.5066474 -1.1307032
#> [13,] -0.88698140  1.2835257  2.70007553 -1.62683955  1.7373454  0.7674151
#> [14,]  0.09822294  0.1354290  0.08295489 -1.90434257 -0.6959686 -0.4758358
#> [15,] -1.73137735 -1.0144880 -0.13538960 -0.78353651 -2.2604783 -0.5990277
#> [16,] -0.11722237  0.9009056 -0.24143582  0.46886597 -0.2268108 -0.7730494
#> [17,]  0.60454912 -0.5100643 -0.10603116  0.51091310 -1.6124721 -1.5609728
#> [18,] -0.03073198  2.0264705  0.96111774  0.21029363  0.7546085  0.9171429
#> [19,] -0.87127508 -0.1671697 -1.34406381  1.21762606  0.6537402 -0.9956640
#> [20,]  0.72534165 -0.3026678  0.44312183  0.30510925  0.9028807  1.0932510
#>             [,37]       [,38]      [,39]       [,40]        [,41]       [,42]
#>  [1,] -0.03573185  0.18852647  1.1850998  0.51723342  0.900386190  0.13951074
#>  [2,]  1.17550207  0.32487559 -0.3382983 -0.67171439 -1.031103920  0.59516266
#>  [3,] -0.30417810  0.49906360 -0.2555086  0.09870359  0.886841691  0.16621798
#>  [4,]  0.67119044 -0.08810252 -0.4473719 -0.38641732 -0.212342508 -1.74279474
#>  [5,] -3.01216378 -1.94761831  0.8844479  0.74368562  0.390278112 -0.98465965
#>  [6,]  1.18415313 -0.41026995  0.7108655 -0.72647148 -0.020917808  0.64817889
#>  [7,] -2.00852306  0.42073725 -0.4885611 -0.27855014 -0.927511490  0.02263407
#>  [8,]  1.20569188  1.40434402 -0.7092102  0.62933831  0.263026128  0.02343003
#>  [9,] -0.04637141  1.17347625 -0.7273762  0.53907168  0.745839229  0.78533003
#> [10,]  0.72280702 -0.31404467  0.2156488 -1.14947172 -0.908578248 -0.55308635
#> [11,] -0.35953706 -1.44623156 -1.1882553 -0.08833947  0.494824157 -1.17918952
#> [12,] -0.37147741  0.45617734 -0.8528608 -0.04642290 -0.309679482  0.21257287
#> [13,] -0.19502166 -0.36675121 -0.5379424 -0.74575751 -0.021931462  1.91079078
#> [14,] -0.53693845  0.33838959  0.2581533  0.52117867  1.109309382  0.45431813
#> [15,] -1.51447014 -0.60779623  0.1193512 -1.56007845  0.588022326 -0.53800149
#> [16,] -0.32514674  1.15889988 -0.7346078 -2.12506858  0.805825697 -1.16823914
#> [17,] -0.89203305  0.31987138  1.3181781  1.08295653  0.756491074  0.88427643
#> [18,] -0.18823391  2.41074643  0.8157023  0.76474257 -0.001020394  0.83818899
#> [19,] -0.86270649  1.09984767  1.4430806 -2.34105600 -0.494026273  0.14730964
#> [20,]  0.55195415 -0.40603663  0.2490262  0.20630527 -0.733315629 -1.00842736
#>              [,43]       [,44]       [,45]      [,46]       [,47]       [,48]
#>  [1,] -0.878529013 -0.17841821  0.85489041  0.1647993  1.28448665  0.88676929
#>  [2,] -0.004982272  2.08522019  0.20077198  2.0552352 -0.26413749  0.02439772
#>  [3,]  0.644437571  0.69622907  1.00044600  0.6936577  0.41740332  0.22548576
#>  [4,]  1.117478705 -0.74273340 -0.49200271  1.5791488  0.84133902 -0.30575887
#>  [5,] -0.629098503  0.52179721 -1.18411791  0.1333243  1.15985361  0.54483272
#>  [6,]  0.785627605 -0.03197743 -0.12007144 -0.1536998  0.26389767 -1.21807626
#>  [7,] -0.098200190 -0.44782659 -1.54110256 -0.7394632  0.49159026  1.17625073
#>  [8,]  0.308572635 -0.28662517  0.56795641  1.2985011  1.32474112  0.72385538
#>  [9,]  0.684217699  0.40540060 -0.21173300 -0.3975188  1.10200872 -0.31357524
#> [10,]  1.239982360  1.07514824  1.30102708 -1.6377194 -0.02297772 -1.45760673
#> [11,] -1.384686031 -0.73696442  0.17035257  0.2571586  0.56857043  0.50537669
#> [12,]  0.046516237  0.31315679  0.81282611  1.4602595 -0.55665016  1.87782555
#> [13,]  1.554945586  0.25574801 -0.69077737 -0.2801739 -0.36391709  1.53970130
#> [14,]  1.526060845  0.20904165  1.02366761  0.6712802  1.28688659 -0.38956188
#> [15,] -0.814341342 -0.28493350  1.69772074  1.0079676  0.13695637  1.49100271
#> [16,] -1.390816142  1.25636913  0.54234783  0.5681053 -0.61268068  0.05382148
#> [17,] -0.843631086 -0.03112132 -0.07268996  1.1552196  0.72427543  0.24368948
#> [18,]  0.241494040 -1.48071872  2.19358676 -1.6210539 -0.94602200 -0.64335790
#> [19,]  1.024927426 -0.48585889 -0.71760886 -0.2027949  1.53998161 -0.41531433
#> [20,]  0.115703376 -0.13508808 -0.41651444 -0.3214095  0.98699233 -2.03517276
#>            [,49]       [,50]      [,51]       [,52]        [,53]      [,54]
#>  [1,]  0.3522661  0.21107589 -1.2090489 -0.83156999  0.637652417  1.1741418
#>  [2,] -0.8216206  1.06133462  1.2967932 -0.22438849  0.236748072 -2.0365253
#>  [3,]  0.8237898  0.75482317  1.0524069 -1.49970430 -0.324382173 -0.4914289
#>  [4,]  0.4162982 -0.34039207 -0.7261675 -0.61870283  1.530732717 -2.2560801
#>  [5,]  0.5977866  1.78010975 -0.1227392  1.74781460  1.172041638 -1.0380622
#>  [6,] -0.7809834  0.28834419 -0.5687760 -0.07823393 -0.217537916 -0.0772461
#>  [7,]  1.2956375  0.66594768  0.7485877  0.65156574  0.007817514  1.0192710
#>  [8,] -0.4335073 -0.22298321 -0.3801552  1.09015815 -1.054728832 -2.9467165
#>  [9,] -0.5780662 -0.01073382  1.5200717  1.04936925 -1.795849546 -1.6064186
#> [10,]  0.6717012  0.83270521 -0.1628784 -0.96696916 -0.221256164  0.4480565
#> [11,]  0.1729750 -1.20105461 -0.1433636 -1.10444518 -0.720545206 -0.3003390
#> [12,]  0.4308321  0.24056483  0.5275333  0.14388983  0.200376494  2.1134122
#> [13,] -0.5811962 -0.64339778 -0.4815867 -1.62538775 -0.999696549  0.2701665
#> [14,] -0.5810773  0.28609258  0.6621468  0.74244673  2.009443116  0.3811075
#> [15,] -0.7825906  0.79729231 -0.9634687 -1.14768833 -0.458177826  1.3547375
#> [16,]  0.5416078  1.05001368 -0.8478035  1.52846789 -0.481925688 -0.6308588
#> [17,] -1.6668292  0.52824827  1.1248102 -0.06809104 -0.628395321 -1.5508260
#> [18,] -0.6951839 -0.19191378 -0.4355499  0.73142331 -1.564949515 -0.8870315
#> [19,] -0.7134732  0.28366974 -0.2154041  1.49553102  0.277907911  1.1261662
#> [20,] -0.5066758 -0.01213930 -1.1701511  0.40374216  1.635590885  0.1131816
#>            [,55]       [,56]       [,57]       [,58]       [,59]       [,60]
#>  [1,] -1.5838715  2.49566783  0.73876936 -0.87709029  2.38851977 -0.62002041
#>  [2,]  0.7360669 -0.53072871  1.12067516 -0.03298339  0.43517009  1.41264439
#>  [3,] -0.5424816  0.51279842  2.66240526 -0.72176278 -0.99001193 -0.24576246
#>  [4,]  1.7297869 -0.80599212  0.85830770 -0.16943443 -0.00388251  0.43217440
#>  [5,]  0.4431161 -1.85770879 -0.59913377  0.50290773 -0.10105417 -0.11160755
#>  [6,]  0.1076756 -1.20284901 -0.13836785 -0.17070012 -0.16636250 -0.49687983
#>  [7,] -0.3598054  0.72738157 -0.45887164 -1.37416464 -1.91414908  0.52087746
#>  [8,]  1.3190846  0.27634234  1.06773801 -1.39529353 -0.26464131  0.12865874
#>  [9,] -0.2613083  0.34749708 -1.54614187  0.03680003 -0.54887437 -0.25817566
#> [10,] -0.9984883  1.52103207 -0.34073649  2.16502581  0.93804375  0.32774925
#> [11,]  0.2877619  0.84246546  0.96703483  0.58620091  0.71389491 -0.07112157
#> [12,]  0.2930571  1.89924335 -0.03915635 -0.40103969 -0.36577739  0.06103350
#> [13,] -1.2278323 -0.03255339 -2.39753544  0.68394351  1.87448873  1.14946740
#> [14,] -0.2508342 -1.05709908 -1.15417235  0.67147496  0.74059649 -2.58177151
#> [15,] -2.6469970 -1.11652692  1.36258039 -0.56875613  0.79465750 -2.43007289
#> [16,] -1.6647369 -0.28095853 -0.96946928  0.64275076 -1.80470122  1.49467610
#> [17,] -1.5120923 -1.00226137  0.21958586  0.27035564 -0.09471368 -1.71044801
#> [18,] -0.6558666 -1.32506561  0.19579576 -0.18658992  0.02410930 -0.04939609
#> [19,]  0.6387166 -1.07992337 -0.65278664 -0.54193700 -1.34705608  0.25852240
#> [20,]  0.8167419 -0.08707742  0.28488377  0.40260968  0.22642452  0.49802129
#>            [,61]        [,62]        [,63]        [,64]       [,65]       [,66]
#>  [1,] -2.0734202  1.268314046  0.589980565  1.150647382  0.15728030 -0.08552748
#>  [2,] -0.9379095 -0.487275186  0.933141193 -0.205438562  1.08359626 -0.99895387
#>  [3,]  0.6894032  1.364005221 -0.454240463  0.470029213  0.77809395  0.56312532
#>  [4,] -1.1742003 -1.421862105 -1.418233056 -1.787473323 -1.63665749  0.20579932
#>  [5,]  0.6531752 -1.842217474 -0.210572124  0.879631239  0.44579281  0.37248589
#>  [6,]  1.3605952 -1.185903183  0.624441674  0.638144724  1.13087171 -0.82783881
#>  [7,]  0.1980498 -0.564817488  0.937064087 -0.923246979 -0.06755019  1.19299163
#>  [8,] -1.9879229  1.501020579  0.743817244  0.706906654  0.61240767  0.16518174
#>  [9,]  1.3995884 -1.659969104 -0.587024900  0.251379345 -1.48049721  0.73483525
#> [10,]  0.1808737  1.218020447 -1.139473919 -0.004725565 -0.88238696  0.42039507
#> [11,]  0.3641842  1.252569834  0.496585989 -1.253740771 -0.20752891 -1.00752315
#> [12,]  0.5778474  1.101621330 -0.180816188 -0.688524922 -0.40244947  1.23780927
#> [13,]  1.6407714 -1.632399936 -0.260201777 -0.155392632 -0.31973262  0.88739264
#> [14,] -1.1017285 -2.685602399  0.007014031  1.381241504  0.87884707 -0.09261388
#> [15,]  1.1352714  1.053663270  1.382226397  0.517320940  1.28908680  0.41874690
#> [16,] -1.1561784  0.741769953 -1.278404907  0.952111614 -1.93909640 -1.75581445
#> [17,] -1.2164404 -2.768705694  0.566921974  0.423064288  0.07278306  0.29806497
#> [18,]  1.1314138  0.005456635 -0.957253179  0.122316551  0.95246510  1.47152734
#> [19,]  0.9009870 -1.781053425  0.074845128  1.158858514 -0.00532533  0.10816840
#> [20,] -0.6119601  0.256339139 -0.910207520  1.907944043  0.81879144  0.04006544
#>              [,67]       [,68]       [,69]      [,70]      [,71]       [,72]
#>  [1,]  0.028508302  1.37832330 -1.65322897 -0.2805187 -1.1876929 -0.16707925
#>  [2,]  0.730770391  0.45941533  0.55333508  0.9523596  1.3632629  1.80025247
#>  [3,]  0.721889354  0.48123808 -0.77341027 -0.2540012  3.8125803 -0.70818084
#>  [4,]  0.441595794  0.94396826 -1.04648803 -0.6056930 -1.7611472  0.81922078
#>  [5,]  0.203508820 -0.56839002 -0.68975150  0.9912493  0.6464894  0.28394300
#>  [6,] -0.861312945 -1.24592590  0.34253659 -2.0764318 -1.3433564  1.32356115
#>  [7,] -0.249387100 -0.39144754  1.67183286 -0.6603536  0.1488688 -0.20854356
#>  [8,]  0.331266844 -0.01127341 -0.96005031 -0.2362137  1.0123609 -1.53210388
#>  [9,]  2.071403671  0.64288621 -0.09604987  0.6932916 -1.0725807  0.04651940
#> [10,] -0.834602773  0.18796458 -0.56234113  1.5384312 -2.0580558  1.18338315
#> [11,] -0.982768991  0.27620613 -0.79600441  1.5272841  0.1776440 -1.22547324
#> [12,] -1.339535295 -0.39243138  0.35892133  1.7082707  1.5898501  0.69127232
#> [13,]  0.242662434 -0.25813305 -0.22474377  0.2369566 -0.2846795  0.49211534
#> [14,] -0.277975908  1.49049279 -0.34730097 -0.2152228  1.4684623  0.25213485
#> [15,]  0.003371298  1.46582983 -1.26017269  1.1738768 -0.7095292  2.22979453
#> [16,] -0.230910497  0.52738054  2.24926361  0.2768507 -0.5226356  0.32266619
#> [17,] -1.551135977 -1.97327590 -1.42640232 -2.4861487 -0.7273097  2.85804162
#> [18,]  0.032008937  1.23738748  0.57982435 -0.4841155 -0.1373628 -1.54889847
#> [19,]  1.222690877  0.92010988 -2.65363714  1.4974247  2.1839913  0.99843840
#> [20,] -0.374327714 -0.14907135 -1.03865936  0.9738209 -0.4203423 -0.06623606
#>             [,73]        [,74]       [,75]       [,76]         [,77]
#>  [1,] -1.61885892 -0.733638076  1.84881803 -0.06482892  0.3359319660
#>  [2,]  0.94836909  0.838675379 -0.13353654 -0.42723063 -1.8384967439
#>  [3,] -0.01577496  0.120209221 -0.71177351  0.59988184  1.5006875766
#>  [4,]  1.77583084 -0.002134996 -0.43683789  0.22766399  1.1275464900
#>  [5,]  0.79259552 -0.016871204  0.77944743 -0.79949363 -0.4173620044
#>  [6,]  1.32185680  1.841161971 -0.39800337 -1.60882706 -1.7506118685
#>  [7,] -1.12697412  1.416525092 -0.86067113 -1.77653585  0.6017462434
#>  [8,]  1.00304780  0.078774519  0.05018479 -1.05374360  0.1709978318
#>  [9,]  0.56601414 -0.550436633 -2.03862806 -1.54898533 -0.4882762341
#> [10,] -0.05853690 -0.280418325 -0.78839537 -0.90683070  0.9577386434
#> [11,] -0.26968873 -2.065285225 -1.70219635  1.39075409  0.7645781846
#> [12,]  1.29314391  0.499512488 -0.30513178 -0.34235315 -0.1927547767
#> [13,]  0.03308461 -2.397153733  1.48832333  1.16911678 -0.9353133746
#> [14,]  1.85900886  1.134052900  0.04667170  2.52616196  0.5238684803
#> [15,]  0.55233389 -0.808755370 -0.15619668 -1.30602474  0.5307647552
#> [16,] -1.33879794 -0.766042846  0.28912439 -0.08736298  1.9591899138
#> [17,] -1.14099145  1.011235742  0.39472252  1.55686080  0.0475953214
#> [18,] -0.79153215  0.170251229  0.46398395 -0.04227645 -1.6901542095
#> [19,]  1.18748892 -1.397233353  2.01993962  0.97514195  0.0009172159
#> [20,]  0.29542597  1.751642296 -0.65414616 -1.17587537 -1.0654711856
#>             [,78]      [,79]      [,80]      [,81]      [,82]       [,83]
#>  [1,] -0.53904035  0.4659718  1.5586541  0.2779890  1.9760054 -0.37601055
#>  [2,] -0.15875216  0.1892923  1.8115396 -0.1185900  0.5128462 -0.09780282
#>  [3,] -0.60687889 -0.9836969  0.2586178 -0.2958575 -0.1408976  1.50023209
#>  [4,]  1.20357036  0.9829257  0.5750962  0.7907331  1.3734946  0.53400556
#>  [5,]  1.20086502  0.1349108  1.5302542  1.3396013  1.0942520 -0.13456260
#>  [6,]  1.32699404  0.1699949  0.4495906  0.6336019  1.0754754  0.81475095
#>  [7,] -0.29857019 -0.1475454  1.1182165  1.1765474 -0.9606134  0.54345427
#>  [8,] -1.44312679 -0.8709408  0.7059092  1.0517861 -0.5846152  0.54070502
#>  [9,] -1.12413450  0.2661394 -1.3253824 -0.7846112  1.6900789  0.70436289
#> [10,]  1.12150987  1.4778066  0.8539253  0.9661272  0.4315164  1.22927523
#> [11,] -0.69630042 -0.4661457  0.1362369 -0.8136528 -0.8184293 -2.05193859
#> [12,] -1.14693313 -0.2709765 -0.8394360  0.3919482  0.3257654  0.14942485
#> [13,]  1.28145975 -1.2709885 -1.4606038  1.2828903 -0.1695890 -1.56321324
#> [14,] -0.07461852 -0.2890526  0.4377567 -0.6963668 -1.2916284 -0.84522620
#> [15,]  1.52984807  1.0844884 -0.3461915  0.7718974 -0.1695980  0.97497947
#> [16,]  1.14177644 -2.0046722  1.6265208  0.5525843  0.2999417  0.74341098
#> [17,] -1.45015013  1.7117634  1.0571177  0.2451662 -1.9073718  2.48899875
#> [18,] -1.15119109  0.5986384 -0.9164453  1.0246344 -0.6977235  0.16779758
#> [19,]  1.01115806  1.0970555 -0.3660224 -0.9619781 -0.1445998 -0.32389041
#> [20,] -0.99629293 -1.2203224  0.9614140 -0.7360079  0.3932049  0.22778990
#>             [,84]      [,85]        [,86]      [,87]       [,88]       [,89]
#>  [1,]  0.31362619  0.4729721  0.481411048  0.3472009  0.15837269 -0.41032884
#>  [2,]  1.49920459 -2.5868960  0.690714442 -0.4700752 -1.11748085 -0.23866430
#>  [3,] -2.60060228 -1.6002552  0.011925567  0.1199367 -0.05741295 -0.43996677
#>  [4,] -0.35029605  0.7870831 -0.491100116 -0.4790227  0.48222679 -0.55736926
#>  [5,]  0.61832189 -0.9310347 -0.180825544 -1.3417446 -0.59255612  1.13245920
#>  [6,]  0.93376776 -1.4429731 -0.368449154 -0.4043311 -1.14011214 -0.13601412
#>  [7,] -0.10984575 -0.1698735 -1.659413927  0.1542336 -0.30816343 -0.85334259
#>  [8,]  0.96922887 -0.6598937  0.465872642 -0.3003545 -0.96160561  0.01803656
#>  [9,]  2.15121875 -0.9012517  0.927548094 -0.6141065  0.17121190  0.82707032
#> [10,]  1.30549596  0.2646120  0.317013122 -0.1163616  0.14931443  0.30521508
#> [11,]  1.11937893  2.6324942 -0.317113090 -1.0598354 -0.49439891  0.51986393
#> [12,] -0.53946860  0.9859009 -0.564275404 -0.3247652  1.05775809 -0.20906883
#> [13,]  0.01886957  1.0590676  2.218552132  0.8062420  0.10635586 -0.08435526
#> [14,]  1.22123014 -0.9404325  1.405653557 -0.7573788 -2.25213853  0.68212941
#> [15,] -1.55348218 -0.2820139  0.606604576 -0.6228313 -0.32307787  1.38897448
#> [16,] -0.83521285 -0.1928180  1.300060004 -0.8821303  1.42855457 -0.92634065
#> [17,] -1.69874344  0.3402877 -0.490009407 -0.1871107 -0.10449867 -0.82857438
#> [18,]  0.78679264 -1.0019545 -0.345046720 -0.2861776  0.64666889  1.97800084
#> [19,]  1.54608127  1.0398827 -0.001244959 -0.7817166  0.33149909  1.28857698
#> [20,]  0.78884795 -1.7263249 -1.807893834 -3.2294978  0.05868121 -0.41070389
#>             [,90]       [,91]       [,92]       [,93]      [,94]        [,95]
#>  [1,]  0.37503131  1.61820620  1.97636957  0.58252465  0.2764916  1.913018306
#>  [2,]  1.68880046  1.30546811 -0.36870639 -1.40436258 -1.3058684 -0.232222142
#>  [3,] -0.30600009 -0.53001239 -0.21514571 -0.02621984  0.9272744 -1.634745937
#>  [4,]  0.66746285 -1.77514362 -1.05238833 -0.26631748 -1.5781636 -0.142129123
#>  [5,]  0.13508351 -0.28112064 -2.80977693 -0.57132785  1.0064043 -0.520068655
#>  [6,] -0.39172740 -0.36898204  0.65673797  0.78894210  1.4574005 -2.059410465
#>  [7,]  0.31867639  0.29958872 -0.66207671 -1.84507669  0.1831583 -1.200988850
#>  [8,]  0.59938390  1.56920528  0.83044282 -0.74606534 -1.5207625  0.624382125
#>  [9,]  0.26577645 -0.44710713 -0.27059239 -2.95264418  1.3671977  0.337456367
#> [10,]  0.64772237  0.89598493 -2.11908740 -0.68469910  0.1434344  0.784041802
#> [11,]  0.30138157  0.75030695  1.36322472 -0.17487121 -0.3643946 -1.086133090
#> [12,]  0.80898961  0.01196499  2.06309243  1.00904032 -0.4486712  1.953897880
#> [13,]  1.20174982  0.87365528 -1.32829245 -0.47044875 -0.2177144 -1.664211487
#> [14,]  0.35769140 -1.27668677  0.94045377 -0.08625320  0.8001145 -1.541071179
#> [15,] -1.93771441  1.17811571  0.04814796 -1.19749723 -0.6465730  0.598097226
#> [16,] -0.04798623 -0.15355582  0.22704760 -1.87511980 -0.1469423  0.536162555
#> [17,]  0.73676465 -0.09476432  0.20639926 -0.25918579 -0.7328225 -0.277782081
#> [18,] -1.20907906  0.32428986  0.59072691  0.16819156  1.6006736 -0.003576539
#> [19,]  0.81585596 -2.46809606 -0.18364356 -0.43824664 -1.3453615  0.408138387
#> [20,] -0.07292253 -0.37546062  2.75762831 -0.70746344 -0.5033114  0.059936863
#>            [,96]        [,97]       [,98]      [,99]      [,100]
#>  [1,] -1.0601106  0.006117106  1.94443658 -0.3390718 -0.13809053
#>  [2,] -0.6275301 -0.579790917 -0.04583145  1.1676814  0.07545568
#>  [3,]  1.0079855 -0.597600967 -0.85205180 -0.2248013 -0.24707548
#>  [4,] -0.8664629 -0.250382500 -1.59514588 -1.2904141 -1.35037031
#>  [5,]  0.1088901  1.538428030 -0.23534841 -0.6603683  1.03316592
#>  [6,]  1.3605890  2.387248345  0.10640793 -0.7495695 -1.06624865
#>  [7,]  1.0587025 -1.262633209 -0.72247250  0.9601573 -0.74152299
#>  [8,]  0.2805595  0.267957668  0.54770152  0.7152799 -0.02769932
#>  [9,]  0.2687907 -0.041980038 -1.50162995 -1.0800344 -0.22655761
#> [10,]  1.5123061  0.748563645  0.15707401  0.1567482 -0.93874446
#> [11,] -0.3796059 -0.837117598  0.26033681 -0.1263609 -1.41954722
#> [12,]  0.1878553 -0.350966692  0.13568016 -0.7307331 -0.57386341
#> [13,]  0.6216689  0.303181171  0.74703313 -1.1349887 -0.70069293
#> [14,]  1.3704532  1.433419144  0.78940183 -0.3635818  0.75808558
#> [15,] -0.6551301 -0.923045572  0.06713694 -0.6634514 -0.33954911
#> [16,] -0.1685582  0.087861663  1.51929526 -0.6271574 -0.30786498
#> [17,]  0.8491342  0.110296754 -1.04740247  0.4924049  0.13842693
#> [18,]  0.9579069 -1.063104165  1.08115505 -1.3074660  1.22176700
#> [19,]  0.9128196  0.182933043  0.22960237  0.9076349  1.04287489
#> [20,] -1.7163911  2.124747579  0.09297220  0.2102978 -2.46870320
#> 
#> $id
#>  [1]  1  1  2  2  3  3  4  4  5  5  6  6  7  7  8  8  9  9 10 10
#> 
#> $visit
#>  [1] 1 2 1 2 1 2 1 2 1 2 1 2 1 2 1 2 1 2 1 2
#> 
#> $n
#> [1] 20
#> 
#> $p
#> [1] 100
#> 
#> $I
#> [1] 10
#>